Ref #258 Moved parsing javacore data to Javacore class
